requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal ==1.2.0
async-timeout   ==        4.0.2
asynctest       ==        0.13.0
attrs           ==        22.2.0
cachetools    ==          4.2.4
certifi            ==     2022.12.7
charset-normalizer    ==  2.0.12
colorama       ==         0.4.5
config         ==         0.5.1
contourpy   ==            1.0.7
cycler            ==      0.11.0
dataclasses     ==        0.8
decorator        ==       4.4.2
einops            ==      0.4.1
filelock          ==      3.11.0
fonttools       ==        4.39.3
frozenlist       ==       1.2.0
fsspec           ==       2022.1.0
future           ==       0.18.3
google-auth    ==         2.16.2
google-auth-oauthlib  ==  0.4.6
grad-cam     ==           1.4.6
grpcio            ==      1.48.2
huggingface-hub     ==    0.13.4
idna              ==      3.4
idna-ssl        ==        1.1.0
imageio         ==        2.15.0
importlib-metadata   ==   4.8.3
importlib-resources   ==  5.4.0
Jinja2             ==     3.0.3
joblib            ==      1.1.1
jsonpatch     ==          1.32
jsonpointer     ==        2.3
keras            ==       2.10.0
kiwisolver       ==       1.4.4
lazy_loader     ==        0.2
lion-pytorch       ==     0.0.7
Markdown          ==      3.3.7
MarkupSafe     ==         2.0.1
matplotlib         ==     3.7.1
mpmath           ==       1.3.0
multidict          ==     5.2.0
networkx      ==          3.1
numpy           ==        1.24.3
nystrom-attention   ==    0.0.11
oauthlib         ==       3.2.2
opencv-python     ==      4.7.0.72
openslide-python  ==      1.1.2
packaging    ==           21.3
pandas         ==         2.0.1
Pillow           ==       9.5.0
plotly          ==        5.13.1
protobuf      ==          3.19.6
pyasn1         ==         0.4.8
pyasn1-modules     ==     0.2.8
pycparser    ==           2.21
pyDeprecate    ==         0.3.1
pydicom           ==      2.3.1
pyparsing        ==       3.0.9
pypiwin32        ==       223
python-dateutil     ==    2.8.2
pytorch-lightning    ==   1.5.10
pytz         ==           2023.3
PyWavelets     ==         1.1.1
pywin32         ==        306
PyYAML        ==          6.0
requests        ==        2.27.1
requests-oauthlib    ==   1.3.1
rsa           ==          4.9
scikit-image      ==      0.20.0
scikit-lear         ==    0.1
scikit-learn      ==      1.2.2
scipy           ==        1.9.1
setuptools        ==      59.5.0
six             ==        1.16.0
sklearn         ==        0.0.post1
sympy             ==      1.11.1
tenacity          ==      8.2.2
tensorboard          ==   2.10.1
tensorboard-data-server ==0.6.1
tensorboard-plugin-wit  ==1.8.1
tensorboardX         ==   2.6
threadpoolctl     ==      3.1.0
tifffile           ==     2020.9.3
timm          ==          0.6.13
tornado         ==        6.1
tqdm             ==       4.64.1
ttach            ==       0.0.3
typing_extensions     ==  4.1.1
tzdata          ==        2023.3
urllib3               ==  1.26.14
websocket-client    ==    1.3.1
Werkzeug       ==         2.0.3
wheel             ==      0.37.1
wincertstore       ==    0.2
yarl              ==      1.7.2
zipp             ==       3.6.0